Convert BindingAnnotation to TS
Update target ES version to support accessor properties
